+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to WireTAP will be documented in this file.
 
-## [Unreleased]
+## [0.8.0] - 2026-06-17
 
 ### Added